Add Color shade for the implementation in the code

+++ src/main/java/math/Normeprod.java
@@ -4,7 +4,7 @@ import graph.Matrice;
 
 public class Normeprod {
 	protected ProduitScalaire produitScalaire;
-	protected int racineProdScalaire;
+	protected double racineProdScalaire;
 	protected Matrice matrice; 
 	private double dx;
 	private double dy;
@@ -15,7 +15,7 @@ public class Normeprod {
 		dx = produitScalaire.getMatrice().getX(0);
 		dy = produitScalaire.getMatrice().getY(0);
 		dz = produitScalaire.getMatrice().getZ(0);
-		racineProdScalaire = (int) Math.sqrt(dx*dx + dy*dy +dz*dz);
+		racineProdScalaire = Math.sqrt(dx*dx + dy*dy +dz*dz);
 		matrice = new Matrice(1,1);
 		matrice.add(dx/racineProdScalaire, dy/racineProdScalaire, dz/racineProdScalaire, 1);
 	}
